+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## CURRENT
 - Fixes tooltip closing with option `followPointer: true` when mouse is over tooltip
+- Adds classes `closing` and `opening` on component during animations
 - Adds method `Closable.closeAfterDelay(delay, callback)`
+- Adds method `Closable.isClosing()`
+- Adds method `Closable.isOpening()`
 
 ## v0.13.1
 - Fixes application of CSS styles with dash in method `Cuic.css()`